hygiene: delegate generate_new_world and auto_save_on_shutdown out of main.rs
Wire up the orphaned app::world_generator module and add app::autosave so
App::generate_new_world and App::auto_save_on_shutdown are thin delegators
rather than 120-line inline implementations. main.rs shrinks from 831 to 653
lines; no behaviour changes.
